python-algorithms/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.9.0
openai>=1.0.0
python-dotenv>=1.0.0
requests>=2.31.0
numpy>=1.24.0
tenacity>=8.2.0
httpx>=0.25.0

# 性能测试依赖
psutil>=5.9.0
matplotlib>=3.7.0

